implement suggestions from code review
