+++ xclim/core/units.py
@@ -11,7 +11,6 @@ import re
 import warnings
 from inspect import signature
 from typing import Any, Callable, Optional, Tuple, Union
-from warnings import warn
 
 import pint.converters
 import pint.unit
@@ -34,7 +33,6 @@ __all__ = [
     "to_agg_units",
     "units",
     "units2pint",
-    "check_same_units_and_convert",
 ]
 
 
@@ -710,20 +708,3 @@ def declare_units(
         return wrapper
 
     return dec
-
-
-def check_same_units_and_convert(func) -> Callable:
-    def _check_same_units(sim, ref, **kwargs):
-        units_sim = units2pint(sim.units)
-        units_ref = units2pint(ref.units)
-
-        if units_sim != units_ref:
-            warn(
-                f" sim({units_sim}) and ref({units_ref}) don't have the same units."
-                f" sim will be converted to {units_ref}."
-            )
-            sim = convert_units_to(sim, ref)
-        out = func(sim, ref)
-        return out
-
-    return _check_same_units
